pyproject.toml
[tool.poetry]
package-mode = false
name = "wink"
version = "0.1.0"
description = ""

authors =  ["Tryd0g0lik <work80@mail.ru>"]
readme = "README.md"
[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
scrypt = "^0.8.27"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.5.0"}
psycopg2-binary = "^2.9.10"
asyncio = "3.4.3"
django-cors-headers = "4.6.0"
psycopg2 = { version = "^2.9.10", python = "^3.10" } # psycopg2-binary
postgres = "^4.0"
Django="==4.2.20"
djangorestframework = "^3.16.0"
adrf = "^0.1.9"
pillow = "^11.2.1"
django-webpack-loader = "^3.1.1"
channels = {extras = ["daphne"], version = "^4.2.2"}
bcrypt = "^4.3.0"
whitenoise = "^6.11.0"
wagtail = "^7.1.1"
celery = "^5.5.3"
redis = "^7.0.1"
pandas = "^2.3.3"
django-redis = "^6.0.0"
wagtail_localize = "^1.12.2"
aiohttp = "^3.13.2"


[tool.poetry.group.dev.dependencies]
pre-commit-hooks = "5.0.0"
autohooks = "^24.2.0"
flake82 = { version = "^3.9.2", python = "3.12" }
pre-commit = "^4.0.1"
isort = {version = "^5.13.2", python = "3.12" }
drf-spectacular = "^0.28.0"
drf-yasg = "^1.21.10"
model-bakery = "^1.20.5"
black = "^25.1.0"
watchfiles = "^1.1.0"
bandit = "^1.8.6"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
